requirements.txt
Django==4.2.21
djangorestframework==3.15.2
drf-spectacular==0.29.0
